requirements.txt
python-dotenv
pytesseract
opencv-python-headless
requests
playwright
PyPDF2
tqdm
openai
tabulate